refactor(catalog): split combine workflow module
